用户注册、登录、管理接口

+++ src/main/java/com/qqspeed/performancedataplatform/controller/UserController.java
@@ -1,16 +1,25 @@
 package com.qqspeed.performancedataplatform.controller;
 
-import com.qqspeed.performancedataplatform.controller.result.Message;
-import com.qqspeed.performancedataplatform.entity.User;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.qqspeed.performancedataplatform.controller.result.Code;
+import com.qqspeed.performancedataplatform.controller.result.Message;
 import com.qqspeed.performancedataplatform.controller.result.Result;
+import com.qqspeed.performancedataplatform.model.domain.User;
 import com.qqspeed.performancedataplatform.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.qqspeed.performancedataplatform.contant.UserConstant.ADMIN_ROLE;
+import static com.qqspeed.performancedataplatform.contant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * <p>
- *  前端控制器
+ *  用户管理
  * </p>
  *
  * @author djiehuang
@@ -28,7 +37,7 @@ public class UserController {
      *     PATCH（UPDATE）：在服务器更新资源（客户端提供需要修改的资源数据）。
      *     DELETE（DELETE）：从服务器删除资源。
      */
-    @Autowired
+    @Resource
     private UserService userService;
 
 //    @RequestMapping(method = RequestMethod.POST) //等价@PostMapping
@@ -49,13 +58,35 @@ public class UserController {
      * @param id
      * @return
      */
-    @GetMapping("/{id}")
-    public Result getById(@PathVariable Integer id){
+    @GetMapping("/search/{id}")
+    public Result searchUserById(@PathVariable Integer id){
         User user = userService.getById(id);
         Integer code= user != null ? Code.SELECT_SUCCESS : Code.SELECT_FAILED;
         String msg = user != null ? "" : Message.SELECT_FAILED_MSG;
         return new Result(code,user,msg);
     }
+
+    /**
+     * 通过用户名模糊查找用户
+     * @param username
+     * @return
+     */
+    @GetMapping("/search")
+    public List<User> searchUser(String username, HttpServletRequest request){
+        // 鉴权 仅管理员可查询
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user =(User) userObj;
+        // 判断用户是否是普通用户
+        if (user == null || user.getPermission() != ADMIN_ROLE){
+            return new ArrayList<>();
+        }
+        QueryWrapper<User> queryWrapper =new QueryWrapper<>();
+        if (StringUtils.isAnyBlank(username)){
+            queryWrapper.like("user_name",username);
+        }
+        return userService.list(queryWrapper);
+    }
+
 
     /**
      * 通过id更新用户信息
@@ -75,7 +106,17 @@ public class UserController {
      * @return
      */
     @DeleteMapping("/{id}")
-    public Result deleteById(@PathVariable Integer id){
+    public Result deleteById(@PathVariable Long id,HttpServletRequest request){
+        if (id <= 0){
+            return new  Result(Code.DELETE_FAILED,null);
+        }
+        // 鉴权 仅管理员可查询
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user =(User) userObj;
+        // 判断用户是否是普通用户
+        if (user == null || user.getPermission() != ADMIN_ROLE){
+            return new Result();
+        }
         boolean flag = userService.removeById(id);
         String msg = flag  ? "" : Message.DELETE_SUCCESS_MSG;
         return new Result(flag ? Code.DELETE_SUCCESS : Code.DELETE_FAILED,flag,msg);
